refactor: Update icon factory to include additional social icons
This commit updates the icon factory in the `icon.tsx` file to include additional social icons such as Facebook, Github, Instagram, and more. This allows for a wider range of social icons to be used throughout the application. The `getSocialIcons` method has been added to retrieve the list of social icons. This change enhances the flexibility and customization options for social media integration.